requirements.txt
pvporcupine==3.0.5
pvporcupinedemo==3.0.5
PyAudio==0.2.14
loguru==0.7.3
flask==3.1.0
flask-socketio==5.3.4
arcade==3.0.2
groq==0.20.0
numpy==2.2.4
sounddevice==0.5.1
azure-cognitiveservices-speech==1.43.0
dotenv==0.9.9
tiktoken==0.9.0
tavily-python==0.5.1
yt-dlp==2025.2.19
mpv==1.0.7
google-api-python-client==2.165.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
mysql-connector-python==9.2.0
python-socketio==5.7.2
websocket-client==1.6.2
python-dateutil==2.9.0.post0
aiofiles==24.1.0
google-generativeai==0.8.4
requests-toolbelt==1.0.0
paho-mqtt==1.6.1
websockets==15.0.1
openai==1.68.0